+++ lib/qc/modules/qc_actions/widgets/qc_action_widget.dart
@@ -4,7 +4,6 @@ import 'dart:convert';
 import 'package:core/core.dart';
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_trc/qc/modules/d2c_video/screens/d2c_video_home_screen.dart';
 import 'package:flutter_trc/qc/modules/data_wipe/screens/data_wipe_home_screen.dart';
 import 'package:flutter_trc/qc/modules/device_details/screens/device_details_screen.dart';
 import 'package:flutter_trc/qc/modules/device_receive_module/screens/device_receive_screen.dart';
@@ -12,6 +11,7 @@ import 'package:flutter_trc/qc/modules/external_audit/external_audit_home_screen
 import 'package:flutter_trc/qc/modules/gaurd/screens/qc_guard_home_screen.dart';
 import 'package:flutter_trc/qc/modules/imei_validator/resources/imei_qrcode_response.dart';
 import 'package:flutter_trc/qc/modules/imei_validator/screens/imei_validator_screen.dart';
+import 'package:flutter_trc/qc/modules/pixel_testing/widget/pixel_testing_button_widget.dart';
 import 'package:flutter_trc/qc/modules/qc_actions/resources/services.dart';
 import 'package:flutter_trc/qc/modules/re_qc/screens/re_qc_list_screen.dart';
 import 'package:flutter_trc/qc/modules/stock_transfer/screens/stock_transfer_list_screen.dart';
@@ -84,6 +84,8 @@ class _QCActionWidgetState extends State<QCActionWidget> {
           mainAxisAlignment: MainAxisAlignment.center,
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
+            PixelTestingButtonWidget(),
+            const SizedBox(height: Dimens.space_16),
             QcRolePermissionWidget(
               role: QcRole.qcElss,
               child: CshBigButton(