+++ lib/src/modules/rider/pending_delivery/receive/widgets/delivery_receive_list_widget.dart
@@ -1,4 +1,5 @@
 import 'package:components/components.dart';
+import 'package:components/resources/list/list_request.dart';
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_trc/src/modules/rider/l10n.dart';
@@ -20,44 +21,55 @@ class _DeliveryReceiveListWidgetState extends State<DeliveryReceiveListWidget> {
   final CshListController _listController = CshListController();
   late DeliveryReceiveProvider provider;
   late L10n l10;
+  FilterConfig? _filterConfig;
+  var _listRefreshCounter = 0;
+  bool _isListenerAdded = false;
 
   @override
   void didChangeDependencies() {
     provider = Provider.of<DeliveryReceiveProvider>(context, listen: false);
     l10 = L10n(context);
-    provider.addListener(() {
-      _refreshList();
-    });
+    if (!_isListenerAdded) {
+      _isListenerAdded = true;
+      provider.addListener(_refreshList);
+    }
     super.didChangeDependencies();
   }
 
   FilterConfig _getFilterConfig() {
-    return FilterConfig(filterData: [
-      CshFilterData(
-        label: "Search Barcode",
-        field: 'partBarcode',
-        crudFilter: 'barcode',
-        filterType: CshFilterType.input,
-        valueType: CshFilterValueType.contains,
-        position: FilterPosition.top,
-        keyboardType: TextInputType.text,
-        filterGroup: FilterGroupType.multipleTypeSearch,
-      ),
+    return FilterConfig(initialFilter: [
+      if (!Validator.isNullOrEmpty(provider.searchQuery))
+        AdminFilterList(
+          type: CshFilterValueType.contains.value,
+          field: 'partStock.barcode',
+          value: AdminFilterData(search: provider.searchQuery),
+        ),
     ]);
   }
 
   void _refreshList() {
-    _listController.refresh();
+    setState(() {
+      _listRefreshCounter++;
+      _filterConfig = _getFilterConfig();
+    });
+  }
+
+  @override
+  void dispose() {
+    provider.removeListener(_refreshList);
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
+    _filterConfig ??= _getFilterConfig();
     return CshApiList<Part>(
+      key: ObjectKey(_listRefreshCounter),
       apiConfig: ListApiConfig(
         apiUrl: "/rider/delivery/pickup/pending",
         serviceGroup: TRCServiceGroups.unifyTrc,
       ),
-      filterConfig: _getFilterConfig(),
+      filterConfig: _filterConfig,
       controller: _listController,
       itemFromJson: Part.fromJson,
       shimmerLoaderWidget: const CshShimmer(height: Dimens.space_60),
@@ -66,7 +78,7 @@ class _DeliveryReceiveListWidgetState extends State<DeliveryReceiveListWidget> {
       isHideCoreFilterButton: true,
       getRowWidget: (item, index) {
         final data = item;
-      return ItemDeliveryReceiveWidget(
+        return ItemDeliveryReceiveWidget(
           item: data!,
           onReceiveConfirm: _refreshList,
         );